Electrum: represent and persist block heights as 32 bits signed ints
Int.MaxValue is about 40,000 years of block which should be enough, and it will fix the encoding
problem users on testnet when there's a reorg and one of their txs has a height of -1.

Side-note: changing the wallet codec can be done easily: if we cannot read and decode persisted data
we just start with an empty wallet and retrieve all wallet data from electrum servers, and once it's
ready it will be encoded with the new codec and saved.

+++ eclair-core/src/main/scala/fr/acinq/eclair/blockchain/electrum/ElectrumWallet.scala
@@ -669,7 +669,7 @@ object ElectrumWallet {
                   changeKeys: Vector[ExtendedPrivateKey],
                   status: Map[BinaryData, String],
                   transactions: Map[BinaryData, Transaction],
-                  heights: Map[BinaryData, Long],
+                  heights: Map[BinaryData, Int],
                   history: Map[BinaryData, List[ElectrumClient.TransactionHistoryItem]],
                   proofs: Map[BinaryData, GetMerkleResponse],
                   locks: Set[Transaction],
@@ -1004,7 +1004,7 @@ object ElectrumWallet {
             }
             history + (scriptHash -> entry)
         }
-      this.copy(locks = this.locks - tx, transactions = this.transactions + (tx.txid -> tx), heights = this.heights + (tx.txid -> 0L), history = history1)
+      this.copy(locks = this.locks - tx, transactions = this.transactions + (tx.txid -> tx), heights = this.heights + (tx.txid -> 0), history = history1)
     }
 
     /**
@@ -1042,7 +1042,7 @@ object ElectrumWallet {
                             changeKeysCount: Int,
                             status: Map[BinaryData, String],
                             transactions: Map[BinaryData, Transaction],
-                            heights: Map[BinaryData, Long],
+                            heights: Map[BinaryData, Int],
                             history: Map[BinaryData, List[ElectrumClient.TransactionHistoryItem]],
                             proofs: Map[BinaryData, GetMerkleResponse],
                             pendingTransactions: List[Transaction],
